refactor(daemon): extract launch configuration
